requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
structlog
pytest
httpx
passlib[argon2]
python-jose[cryptography]